backend/requirements.txt
amqp==2.6.0
argh==0.26.2
asgiref==3.2.8
billiard==3.6.3.0
boto3==1.14.5
botocore==1.17.5
celery==4.4.5
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
jmespath==0.10.0
kombu==4.6.10
Markdown==3.2.2
orjson==3.1.0
pathtools==0.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
watchdog==0.10.2
